Add and use random seed interface for ancestral
Adds a new `--seed` argument to the ancestral command, following the
existing pattern in the refine command, and passes the user-provided
value to TreeTime's `TreeAnc` class to ensure deterministic outcomes for
random samples. Since this PR changes ancestral's root sequence
inference to be stochastic, this new feature allows us to get the same
"random" result with each run of our functional tests.

Correspondingly, this commit updates the non-VCF functional tests of
ancestral to use a fixed seed. This seed should fix the random behavior
to a specific outcome, but we need to update our expected JSON outputs
to match that outcome here. In this case, the change is an introduction
of a "T" at position 14 in the node_root of the "simple genome" data.