Use random seed for multiallele VCF ancestral test
Updates the functional test of multiallele VCF behavior when one of the
alleles is a "N" character. The multiallele VCF in the original
test (setting sample B to genotype of 30G) caused the inference of
alleles at sites 7 and 14 to change substantially in this PR's new
implementation. This commit stabilizes the test by assigning both
samples A and B to the 30G genotype such that all nodes are inferred to
have a 30G and the root node gets a A30G mutation. It should maintain
the originally intended behavior of the test [1].

As part of this change, I've replaced the dynamically modified VCF used
by this test with a static version of the VCF in version control. This
change simplified test debugging by allowing me to inspect the input
file outside of the cram temporary environment and it also fixes an
issue with the original dynamic VCF where an additional line for site 30
was added but the original line was not being removed.

[1] https://github.com/nextstrain/augur/issues/1380