Fix test data for augur translate tests
Updates SNPs in manually-curated VCF used to test augur translate so
they match the new expected genotype for the "simple genome" data at
position 14. This fixes the vcf.t translate test. Also updates the
expected amino acid translations JSON for the vcf with root mutation
test. As part of this update, I've created a complete copy of the
"truth" aa_muts.json with the root mutation and used this instead of the
version that was previously created by modifying the original aa_muts
file with sed. Although this change duplicates test data, I was able to
debug the test more easily with a complete version of the JSON living
outside of the cram tests.