Removes unsupported pylint `no-space-check` option
As of pylint 2.6, `no-space-check` is not a supported option in the
pylint configuration file [1]. This commit removes the option from the
project-specific pylintrc to silence errors from pylint during
development in emacs. Since pylint is now at version 4 and no one else
on the team has run into this issue, I strongly suspect no one else is
using this pylintrc the same way I am. If folks want to make a case for
removing this file entirely in favor of some other IDE-based linting
they prefer, I'm open to that.

[1] https://pylint.pycqa.org/en/latest/whatsnew/2/2.6/summary.html#other-changes