parse: Add invalid characters to test data
Modifies the existing functional tests for parse by adding invalid
characters (spaces and parentheses) which augur parse converts to either
nothing or underscores, respectively. Since parse only transforms strain
names in the sequence output and not the metadata, this change to the
input data causes the functional test to fail. This commit adds a
parallel copy of the metadata ("other") for use in the test where the id
column is an accession and not the strain name and we don't expect the
strain name to get transformed.