Clarify behavior of --lat-longs
The previous text mentioned built in mappings but did not specify where
to find them or specifics around the overriding behavior.

The link to the GitHub-hosted lat_longs.tsv file associated with the
installation's version is not perfect since the local version of
lat_longs.tsv can be changed (i.e. in a development environment), but
this should work for most users.