Skip CI for PRs on Markdown changes
Skips the CI workflow when we change the README or change log in a PR.
Changes to these files should not produce a different outcome in CI, so
we don't need to waste compute rerunning CI for these files. CI will
still run for these changes when pushed to `master`, since the change
log gets built as part of our documentation.