Hide BiopythonDeprecationWarnings when reading certain sequence files
Biopython 1.85 will show a deprecation warning when using format='fasta'
with files that start with anything but '>'. This will be triggered in 2
cases:

1. when attempting to read GenBank files as FASTA
2. when attempting to read a FASTA file with comments

For (1), this deprecation warning will eventually become a ValueError
which works with the existing try/catch block in ancentral.py to
continue parsing as GenBank. For (2), it should be decided independently
of Biopython if Augur should support FASTA with comments.

In both cases, the warning as-is should not be exposed to Augur users.
It is not triggered when reading files format='fasta-pearson', so this
is the easiest thing to do to maintain behavior of Biopython <1.85.

An alternative workaround is to suppress BiopythonDeprecationWarnings
entirely with

    warnings.simplefilter("ignore", BiopythonDeprecationWarning)

but that is less targeted and can lead to suppression of meaningful
warnings.