Hide BiopythonDeprecationWarnings when reading certain sequence files
Biopython 1.85 will show a deprecation warning when using format='fasta'
with files that start with anything but '>'.

The warning as-is should not be exposed to Augur users. It is not
triggered when reading files with format='fasta-pearson', so this is the
easiest thing to do to maintain behavior of Biopython <1.85.