Hide BiopythonDeprecationWarnings when reading certain sequence files
Biopython 1.85 shows a deprecation warning when using format='fasta'
with files that start with anything but '>'.

The warning as-is should not be exposed to Augur users. It is not
triggered when reading files with format='fasta-pearson', so this is the
easiest thing to do continue accepting such files for Biopython
>=1.85.

This way, Augur users get consistent, backwards compatible behavior no
matter the Biopython version they use.