curate/apply-geolocation-rules: Include the default geolocation rules
`--geolocation-rules` is now optional and acts as custom rules that
have precedence over the default geolocation rules in Augur.

The custom rules are applied first and if none of the custom rules are
applied to a record, then use the default rules. This means that if the
custom rules have wildcards, then all locations that match the wildcards
will no longer use the default rules. This is best demonstrated in the
Cram test included in this commit.