[export] preserve behaviour of author(s), numdate
There are a large number of modifications and special-casing around
'author'/'authors' and 'numdate'/'num_date'.

If author or temporal information is set on the metadata / node-data
JSONs (under "author" or "authors", "numdate" or "num_date") we'll
always export it on nodes in the tree. This means the need to specify
author[s] or numdate/num_date in "metadata columns" is a no-op and we
can simplify things by filtering them all out.

For colorings we preserve the existing behaviour of renaming "authors"
to "author".

Going forward a big simplification would be to move the command-line
parsing into the `read_auspice_config` function and generate a unified
config structure ahead of time, but for now I'm only focusing on the
config JSON.