[refine] add 'date_inferred' boolean to output
This is only exported when using `--timetree` and is intended to be
used by `augur export` to inform Auspice whether a date is inferred or
not. An alternative approach would be for `augur export` to parse the
already exported `raw_date` and decide whether that includes ambiguity
but with the proposed changes in <https://github.com/nextstrain/augur/issues/1304>
this approach is cleaner and less error-prone.

Note that `raw_date_constraint` is None on internal nodes and a list
(of len 2) on terminal nodes with ambiguous dates.