Support ISO time intervals
Previously, the only way an interval could be specified was through an
ambiguous date, which has limited flexibility. Conveniently, the
aniso8601 library provides a function to parse the various ISO interval
formats¹.

Note that this supports time information where all other date format
support and logic is based on date without time information. This is
fine because it is trivial to ignore the time information and have
everything work as expected. It seems far-fetched for now, but maybe
someday we will end up using time information, since date is inherently
ambiguous without at least a time zone.

¹ <https://en.wikipedia.org/wiki/ISO_8601#Time_intervals>