align: Remove insertion sequence print statements
These print statements can flood the workflow logs with long insertion
sequences that make it hard to parse. The same insertions are already
output to the `insertion_csv` so there's no reason they need to be
printed to stdout.

Note I opted to keep the warning logs for an insertion of gaps since
that is not captured in the `insertion_csv`.

Resolves <https://github.com/nextstrain/augur/issues/1737>