Sort unweighted group by columns in output file
Set should be used under the hood, but the order of sets is not
guaranteed and it is best to have a guaranteed order for
reproducibility. A simple alphabetical sort is sufficient.

This became prominent when switching the internal representation of
numeric columns such as 'year' from integer to string. The order of
integers seems to be more consistent with sets compared to strings,
and this was inadvertently relied upon in some functional tests.