Add test for branch lengths with ambiguous columns
Adds a functional test for augur refine's behavior when one or more
columns in the alignment are all Ns. In this case, the columns should
get ignored in the final branch length calculations as ambiguous
positions. This test adds an alternate version of the Zika alignment and
the expected branch lengths JSON where the third column of the alignment
has been changed to all Ns. Since this column originally had a variant
in one sample that would have changed the overall branch lengths, this
change produces different a node data JSON output with a shorter branch
length.