[export] allow multiple config files
The merging approach implemented here is specific to the config files
and differs from other merging approaches we may use in the group.
Specifically, when lists (e.g. 'colorings') are found in multiple
configs we don't replace the first list but rather extend it, and use a
customised approach to choosing when to replace elements in the original
list with an updated version (e.g. for 'colorings' if the 'key' of an
individual coloring matches a previously seen one it will replace it in
the list).