Remove unnecessary condition in min/max date filter
The previous commit clarified that the type of `dates[s]` can be either
`float`, `(float, float)`, or `None`. This caused mypy to flag that
all() cannot be used with some of these values. That's true, but mypy
does not know that `np.isscalar()` narrows the type down to non-scalars.

On a second thought, this condition is not necessary because the
comparison on the next line will work regardless of if `dates[s]` is
`float` or `(float, float)`.