[export] remove `fatal` wrapper function
The behaviour of this function was identified as wanting in review¹ and
since we use it twice (within `export_v2.py`) and call `AugurError`
twice I opted to move all usages to the latter. Note that this changes
the printed message slightly, but it's now more consistent with other
code in the project.

¹ <https://github.com/nextstrain/augur/pull/1756#discussion_r2014778621>