[export] use warn wrapper function
The previous code used `warnings.warn` as well as `warning`, where the
latter was a wrapper around the former¹. We now rename the wrapper as
`warn` so that it's consistently used. Note that the only actual usage
of the previous `warning` function was removed in an earlier commit in
this series.

This change was suggested in both the PR description and by others in
review².

¹ The underlying warning category is unchanged (`UserWarning`) but we
now have a stacklevel of 2 not 1. In the current usage I don't observe
any differences.

² <https://github.com/nextstrain/augur/pull/1756#discussion_r2014784934>