Relax search pattern for 'year-month[-day]' dates
Prior to "Use standard date parsing function in subsampling" (041412c5),
this would simply split a date string on `-` to get numbers
corresponding to year/month/day. While there was no explicit
documentation of support for this format, which is a superset of the
explicitly supported ISO 8601 format, it was supported in practice.
041412c5 broke support for that format. It has been restored for
backwards compatibility in most situations (see below for caveats).

There are still some restrictions of the new relaxed search pattern that
did not exist with the former string splitting:

1. Date parts must be within valid date bounds. Prior to 041412c5, date
   parts in values such as '2010-13' were simply treated categorically
   with month=13. The values are now run through a date conversion
   function which will fail to parse out of bounds values, or in the
   case of AmbiguousDate used in <year>-<month>, auto-convert them to
   the closest in-bound value.
2. For <year>-<month>, year must be at most 4 digits, and month at most
   2 digits. Prior to 041412c5, values such as `2014-001` mapped to a
   categorical month=1. Values are now subject to maximum digits imposed
   by AmbiguousDate.regex. We could relax this regex as well, but I
   think it's better to limit rather expand support for these
   nonstandard values.