Cache get_year_month_day
The shift to using this function during date inference for subsampling
(041412c5) provides a good opportunity for a caching speedup: the
function is applied to every date value in the metadata, and dates are
likely to be highly repetitive in large datasets.

I tested this on the current full GISAID metadata which has ~17 million
sequences. Time spent running that function drops from 8 minutes to 0.03
seconds.