Store only the identifier in the priority queue
All metadata for a row was required back when metadata output was based
on information stored in memory. That is no longer the case as of
"filter: Stream filtered metadata to outputs directly" (982bc1c).

This may not improve run time since .loc using a unique index should be
O(1),¹ but it shouldn't hurt.

¹ <https://stackoverflow.com/a/16629125>