Repurpose validate_sequence_file to read_sequence_ids
This will be useful in augur filter, where the return value can be used
directly for observed_sequence_strains.

Note that this walks back on the avoidance of reading a large list of
ids into memory. I haven't dug into how SeqKit implements duplicate
detection, but it must be storing some representation of unique ids
either somewhere in memory or on disk to determine if an id is not
present in the rest of the file. So, instead of offloading the job to
SeqKit, we do this ourselves and gain access to a variable that
represents the ids present in the sequence file. Although it's not
great, there is already precedent for storing this information in memory
with the current implementation of augur filter.