Skip building sequence index if there are no sequence filters
Building a sequence index is costly and takes up the bulk of the time
when working with large datasets. Instead of relying on it to filter
metadata rows based on whether a sequence is available, use the ids from
the sequence file directly since those are computed to check for
duplicates already.

This adjusts the order of operations so that sequence_strains is
determined upfront from --sequences directly, rather than first with
--sequence-index then updated upon inspection of --sequences. This also
means the warning of "Metadata and strain output may not match sequence
output" can be removed.