Make XXXX-XX-XX represent (-∞,∞)
This is the value used by augur curate format-dates to represent a lack
of date information. Conceptually, this should be represent a maximally
wide date interval.

Previously, this value went through AmbiguousDate.range() which returned
`[0001-01-01, <today>]`. The lower end is a limitation of the ISO 8601
format, and the upper limit is a somewhat opinionated default behavior.
Since dates are stored numerically, the maximally wide date interval is
(-∞,∞).

The motivation for making this change is for the upcoming
apply-date-bounds command which will need to check whether ends of an
interval are defined. It doesn't make sense to check against
`0001-01-01` or `<today>` since those are values that can be used to
define an interval.