Parameterize SeqKit's thread usage
Added to filter and merge commands given their internal usage of SeqKit.
Usage was previously hardcoded to the default of using 4 threads, which
can lead to oversubscription in the common approach of running multiple
augur filter calls simultaneously using Snakemake.

The new options are named --nthreads for consistency with align and
tree commands.