Add option to skip first pass of sequence file
This can be useful when working with large datasets such as ncov. Right
now, this is analogous to augur merge's
--skip-input-sequences-validation, but I can see something similar being
added in the future for metadata where it would also make sense to skip
a first pass. The name --skip-checks seems generic enough to cover both
use cases.

I initially¹ thought of reverting back to getting the list of ids and
filtering in a single pass of the FASTA file, however this would require
building the sequence index where not necessary, which is the reason for
the split into multiple passes. It seems better to avoid unnecessarily
building the sequence index which can be costly on large datasets, so
the option added here is a user-customizable workaround.

¹ <https://github.com/nextstrain/augur/commit/d00fb57bbf6b1860db80f453379d5376b71eb586>