Add option to skip first pass of sequence file
This can be useful when working with large datasets such as ncov. Right
now, this is analogous to augur merge's
--skip-input-sequences-validation, but I can see something similar being
added in the future for metadata where it could also make sense to skip
a first pass. The name --skip-checks seems generic enough to cover both
use cases.

I initially thought of reverting back to using a single pass of the
FASTA file to both get the list of ids and subset the file (and
attempted an implementation that still uses SeqKit¹), however that
requires building the sequence index where not necessary, which was the
reason for splitting into multiple passes in the first place.² Building
the sequence index is more costly than read_sequence_ids (seqkit
fx2tab), so I think it makes the most sense to automatically minimize
the burden of the former and add an option to minimize the burden of the
latter.

¹ <https://github.com/nextstrain/augur/commit/d00fb57bbf6b1860db80f453379d5376b71eb586>
² "Skip building sequence index if there are no sequence filters" (37884dba)