Add a profiling mode controlled by AUGUR_PROFILE
I originally thought of making this a top-level option (e.g. augur
--profile), however doing so would lose the ability to profile argument
parsing and everything that happens before that. Given that the slowness
in Augur's startup time is due to those steps, profiling them is
valuable.

Note that this works with commands such as filter and merge which invoke
read-file and write-file because, while calls to the latter will save
profiling results to the path given in the shared env var, they will be
overridden when the parent process saves its profiling results to the
same path.