Default to using a single thread
SeqKit's default of 4 threads works well for a single command
invocation, however it is prone to oversubscription of resources in the
common use case of having a workflow manager run multiple invocations
simultaneously. Multiple threads are to be used at the user's
discretion, where they must be aware of how to do this properly such as
with Snakemake's threads directive.