Vectorize string concatenation
All the operations done in the get_year_month function have a vectorized
equivalent in pandas which is significantly faster.

In local testing of the ncov/open/100k dataset with --group-by month,
this change brings the statement's run time from 11 seconds down to .16
seconds.

Side note: the addition of caching to this function in "Cache other date
functions" (361cb441) was short-sighted. Caching has little benefit with
simple operations such as string concatenation.