Update descriptions for rendering in reST
This improves rendering on the augur subsample documentation page.

Notably, reStructuredText syntax was added to make the link in the query
description a clickable link. With the way help CLI help text is
rendered, it means that reST syntax show up in the augur filter help
text. Not ideal, but seems fine. I used the format where the link target
is placed at the end of the description to reduce clutter.