read-file/write-file: Add supported compression formats to docs
I didn't want to have a link to augur.io.open_file and then link to xopen since
that seem like implementation details that are not relevant to the end user. 
Just be clear and list out the supported compression formats.

Motivated by wanting docs for the pathogen-repo-guide changes in 
https://github.com/nextstrain/pathogen-repo-guide/pull/91