Support defaults
This can help reduce duplication across samples.

I chose to only allow a subset of options to be set as default,
specifically just the selection options (not subsampling). The primary
reason is that there are two mutually exclusive options to configure
sample size: sequences_per_group and max_sequences. While it's possible
to handle cases where one is set in defaults and the other is set in a
sample, I figured it's easier to simply not allow that, at least in the
initial release.