Consider XXXX-XX-XX as missing date
Previously, this was considered an ambiguous date with a range of
'[0001-01-01, present]'. That may be somewhat correct, but it has the
side effect of passing min/max date filters. Other missing dates such as
empty string are dropped by those date filters, and it seems more
reasonable to consider XXXX-XX-XX the same way.