Add option to keep index as a column
pandas's read_csv() converts 'index_col' from a column to the index.
Downstream code can benefit from keeping it as a column in addition to
setting as the index.

There is existing code that relies on the absence of this column (due to
the lack of this feature), so put it behind a function parameter with
default behavior unchanged.

A generic helper function to achieve this for both the chunked and
DataFrame return options of read_csv().