Add CLI option for custom search paths
Previously, the only search path was implicitly the current working
directory. Make this explicit and add an option to search other paths
before it.

This means the notes in option descriptions mentioning cwd as search
path are no longer accurate for augur subsample. Since the descriptions
are shared with augur filter which does not allow custom search paths, I
figured it'd be best to remove the notes entirely.