Replace iterrows() with itertuples()
There was a type error with the previous code:

    subsample.py:353:12 - error: Invalid conditional operand of type "Series[bool]"
        Method __bool__ for type "Series[bool]" returns type "NoReturn" rather than "bool" (reportGeneralTypeIssues)

This is because iterrows() yields Series, and Series.__getitem__ returns
Series[Any] | Any (scalar). Pyright sees row[…] as that type, which
means the outcome of the comparison operator is Series[bool] | bool.
Series[bool] cannot be used as-is for the if condition.

This could be addressed by casting 'int(…) < int(…)', but switching to
itertuples() seems more appropriate as it addresses the type error and
is more efficient.¹

The internal column names have been stripped of their leading
underscores to avoid being renamed.²

¹ https://stackoverflow.com/a/70227993
² "Notes" section in https://pandas.pydata.org/pandas-docs/version/2.3/reference/api/pandas.DataFrame.itertuples.html