[curate apply-geolocation-rules] allow '?' values
Allows '?' to be used interchangeably with the empty value for the
purposes of geographic remappings. See the added tests for real-life
occurrences which this commit fixes.

In our default rules files there are n=5181 rules where only the final
value (location) is empty and these rules can now apply to metadata
values which set "location='?'". There are 44 rules with empty division
& location and one rule (!) with empty country, division & location.